WIP expanded name list in entities
Hi chaps,

I was thinking yesterday about the various developments we are doing
(or have done) around addressing things in the charm store. It occurs
to me that a lot of what we're doing is providing different ways to
address entities. The complexity is creeping in by providing a number
of heuristics for determining what entity is being addressed when an
entity is asked for. My suggestion is to turn the problem on its head
and have the entity contain the list of names by which it may be
addressed. charmstore can then effectively do simple searches for what
is asked for.

Entities can be referenced by names that can refer only to a
particular entity (those with a revision number) or names that can
apply to multiple entities (those without a revision number).

I there propose adding two arrays to the entity document:

ResolvedNames []*charm.Reference UnresolvedNames []UnresolvedName

where UnresolvedName is URL *charm.Reference Series string Revision
int

entities query then becomes as simple as func (s *Store)
EntitiesQuery(url *charm.Reference) *mgo.Query {
    entities := s.DB.Entities() if url.Revision == -1 {
	return entities.Find(bson.D{{"unresolvednames.url", url}})
    } return entities.Find(bson.D{{"resolvednames", url}})
}

publishing and promulgation and any other aliasing typee ideas we
dream up in the future can then be implemented by manipulating the
list of names associated with the entities.

There are other details, but it mostly all follows on from making sure
the entity is stored correctly in mongodb.

I'd appreciate any thoughts.

Thanks, Martin